Move marimo_utils into larch_cli_wrapper
